(feat) Split App into Home-Auth-EventList structure
