+++ client/components/EventCreate.js
@@ -1,0 +1,1 @@
+// EventCreate component
