+++ client/components/EventDetails.js
@@ -1,0 +1,1 @@
+// EventDetails component
