+++ client/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Router, Route, hashHistory } from 'react-router';
 import {Provider} from 'react-redux';
 import {createStore, applyMiddleware} from 'redux';
 import ReduxPromise from 'redux-promise';
 
 import reducers from './reducers';
-import App from './components/App'
+import Home from './components/Home'
 
 const createStoreWithMiddleware = applyMiddleware(ReduxPromise)(createStore);
 
@@ -13,7 +14,7 @@ class Index extends React.Component {
   render() {
     return(
       <div>
-        <App />
+        <Home />
       </div>
     );
   }
