+++ package.json
@@ -33,6 +33,7 @@
     "react-addons-test-utils": "^15.2.1",
     "react-dom": "^15.2.1",
     "react-redux": "^4.4.5",
+    "react-router": "^2.6.1",
     "redux": "^3.5.2",
     "redux-promise": "^0.5.3",
     "socket.io": "^1.4.8",
